修正 podsepc valid 的warning

+++ CJWPayUtils/Util/openssl_wrapper.m
@@ -80,7 +80,7 @@ int rsa_verify_with_public_key_pem(char *message, int message_length
 
 NSString *base64StringFromData(NSData *signature)
 {
-    int signatureLength = [signature length];
+    NSUInteger signatureLength = [signature length];
     unsigned char *outputBuffer = (unsigned char *)malloc(2 * 4 * (signatureLength / 3 + 1));
     int outputLength = EVP_EncodeBlock(outputBuffer, [signature bytes], signatureLength);
     outputBuffer[outputLength] = '\0';
@@ -91,7 +91,7 @@ NSString *base64StringFromData(NSData *signature)
 
 NSData *dataFromBase64String(NSString *base64String)
 {
-    int stringLength = [base64String length];
+    NSUInteger stringLength = [base64String length];
     const unsigned char *strBuffer = (const unsigned char *)[base64String UTF8String];
     unsigned char *outputBuffer = (unsigned char *)malloc(2 * 3 * (stringLength / 4 + 1));
     int outputLength = EVP_DecodeBlock(outputBuffer, strBuffer, stringLength);
@@ -117,7 +117,7 @@ NSData *dataFromBase64String(NSString *base64String)
 NSString *rsaSignString(NSString *stringToSign, NSString *privateKeyFilePath, BOOL *signSuccess)
 {
     const char *message = [stringToSign cStringUsingEncoding:NSUTF8StringEncoding];
-    int messageLength = strlen(message);
+    NSUInteger messageLength = strlen(message);
     unsigned char *sig = (unsigned char *)malloc(256);
     unsigned int sig_len;
     char *filePath = (char *)[privateKeyFilePath cStringUsingEncoding:NSUTF8StringEncoding];
@@ -141,10 +141,10 @@ NSString *rsaSignString(NSString *stringToSign, NSString *privateKeyFilePath, BO
 void rsaVerifyString(NSString *stringToVerify, NSString *signature, NSString *publicKeyFilePath, BOOL *verifySuccess)
 {
     const char *message = [stringToVerify cStringUsingEncoding:NSUTF8StringEncoding];
-    int messageLength = [stringToVerify lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
+    NSUInteger messageLength = [stringToVerify lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
     NSData *signatureData = dataFromBase64String(signature);
     unsigned char *sig = (unsigned char *)[signatureData bytes];
-    unsigned int sig_len = [signatureData length];
+    NSUInteger sig_len = [signatureData length];
     char *filePath = (char *)[publicKeyFilePath cStringUsingEncoding:NSUTF8StringEncoding];
     int verify_ok = rsa_verify_with_public_key_pem((char *)message, messageLength
                                                    , sig, sig_len
@@ -163,7 +163,7 @@ NSString *formattedPEMString(NSString *originalString)
 {    
     NSString *trimmedString = [originalString stringByReplacingOccurrencesOfString:@"\n" withString:@""];
     const char *c = [trimmedString UTF8String];
-    int len = [trimmedString length];
+    NSUInteger len = [trimmedString length];
     NSMutableString *result = [NSMutableString string];
     [result appendString:@"-----BEGIN PRIVATE KEY-----\n"];
     int index = 0;
